Refactor nutrition coach UI components
