+++ client/src/components/meal-planner/coach/NutritionCoachEmptyState.tsx
@@ -1,0 +1,7 @@
+export const NutritionCoachEmptyState = () => (
+  <div className="rounded-xl border border-dashed bg-white/70 p-6 text-sm text-gray-600">
+    All visible coach insights are dismissed for this week. Restore dismissed insights to review them again.
+  </div>
+);
+
+export default NutritionCoachEmptyState;
